Fix datatypes for bit_rate_download and bit_rate_upload

+++ main.go
@@ -15,8 +15,8 @@ var (
 
 type videoBridgeStats struct {
 	Threads                               int     `json:"threads"`
-	BitRateDownload                       int     `json:"bit_rate_download"`
-	BitRateUpload                         int     `json:"bit_rate_upload"`
+	BitRateDownload                       float64 `json:"bit_rate_download"`
+	BitRateUpload                         float64 `json:"bit_rate_upload"`
 	PacketRateDownload                    float64 `json:"packet_rate_download"`
 	PacketRateUpload                      float64 `json:"packet_rate_upload"`
 	LossRateDownload                      float64 `json:"loss_rate_download"`
